Version: 1.0.2 - 2020-04-12 Fixed: 1. Concurrent Modification Exception while clean up banned list. 2. Ban length did not use "Long" after all, until this version.

+++ src/com/jack_l/honeyport/configuration/CachedConfigurationValues.java
@@ -78,7 +78,7 @@ public class CachedConfigurationValues {
      * permanently
      */
     @Builder.Default
-    private int banLength = 0;
+    private long banLength = 0;
 
     /**
      * Fake Server welcome message
